Feature/more perf improvements (#29)
* Made performance and readability tweaks.
* Improved StoreBytes performance for LittleEndianByteIndexer
* Improved performance of BigEndianByteIndexer StoreBytes.
* Added file layout settings.
* Reformatted the code.
* executed performance run.
* removed commented out code.
* Tagged private methods fields and properties with the exclude tag.
* Updated API documentation. 
* Refactored code for readability.
* Changed Bits to a field in BitIndexer and BitMask.
* Corrected which arrays are used for performance tests
* Added more performance benchmarks for non-byte indexer based conversions.
* Removed unused params
* Updated benchmarks to include ReadOnlySpan
* Updated numerous method summary comments for consistency and comprehension.
* Excluded an item from generated documentation.
* Fixed a broken behavior in byte array extensions. (breaking change)
* Refactored some methods to reduce cyclomatic complexity
* Removed conditional compilation for .netstandard2.1/2.0
* Refactored ByteArrayExtensions to match BigEndianByteIndexer behavior of StoreBytes for the ToXYZ methods. Updated API docs.
* Added some explicit conversion operators.

+++ Jcd.BitManipulation.Examples/RunStats.cs
@@ -19,7 +19,7 @@ internal record RunStats
  , double TotalCpuCycles
  , double CpuCyclesPerOperation
  , Quantity<Frequency> AverageCpuFrequency
- , int OperationsPerIteration
+ , long OperationsPerIteration
 )
 {
    public static readonly RunStats Empty = new(0.As(Durations.Second), 0.As(Durations.Second), 0d, 0d, 0.As(Frequencies.Hertz), 0);
